Move already_listening to plugins.util
