+++ docs/api/plugins/util.rst
@@ -1,0 +1,5 @@
+:mod:`letsencrypt.plugins.util`
+-------------------------------
+
+.. automodule:: letsencrypt.plugins.util
+   :members:
